// src/types/user.ts
// export interface ApiUser {
//   id: number;
//   name: string;
//   email: string;
// }

// export interface User {
//   id: number;
//   firstName: string;
//   lastName: string;
//   email: string;
//   department: string;
// }
export interface ApiUser {
  id: number;
  name: string;
  email: string;
}

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  department: string;
}